Created components for each tab
Page file was starting to get pretty long so separated it into components for readability
